requirements.txt
Django==1.10.5
matplotlib==1.5.3
mock==2.0.0
neo4j_driver==1.1.0
networkx==1.11
tweepy==3.5.0
beautifulsoup4==4.5.3
djangorestframework==3.5.3
